refactor: :truck: improve folder structure
