+++ src/renderer/views/Files/subcomponents/FileTreeItem/style.ts
